+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed:
 - Remove autofocus on measurement value field
 - Unify segmented time span controls on dashboard and habit pages
+- Increase analyzed habit completion time span to 90 days
 
 ## [0.8.205] - 2022-12-09
 ### Added: